+++ byu_gui/checkout.py
@@ -1,76 +1,58 @@
+import sys
+import os
 from PyQt4 import QtGui
-import sys, os
 #from byuam.project import Project
+		
+CHECKOUT_WINDOW_WIDTH = 340
+CHECKOUT_WINDOW_HEIGHT = 575
+dept_list = ['Model', 'Rig', 'Animation', 'Layout']
 
-"""
-Checkout Dialog is the Dialog window for the Checkout tool
-"""
-class CheckoutWindow(QtGui.QWidget):
-    def _init_(self):
-        super(CheckoutWindow, self)._init_()
-        
-        # self.project = Project()
-        self.dept_list = ['Model', 'Rig', 'Animation', 'Layout']
+class checkoutWindow(QtGui.QTabWidget):
+	def __init__(self):
+		super(checkoutWindow, self).__init__()
+		self.initUI()
 
-        # Initialize the GUI
-        self.setGeometry(300, 300, 340, 575)
-        self.setWindowTitle('Checkout')
-        #self.setFixedSize(
+	def initUI(self):
+		#define gui elements
+		self.setWindowTitle('Checkout')
+		self.resize(CHECKOUT_WINDOW_WIDTH, CHECKOUT_WINDOW_HEIGHT)
 
-        # Create tabbed view
-        #self.dept_tabs = QtGui.QTabWidget()
-        #for dept in self.dept_list:
-        #    self.dept_tabs.addTab(dept)
+		#create tabs
+		self.dept_tabs = QtGui.QTabWidget()
+		for dept in dept_list:
+			tab = QtGui.QWidget()
+			self.dept_tabs.addTab(tab, dept)
 
-        # Create Label to hold asset info
-        #self.asset_info_label = QtGui.QLabel()
-        #self.asset_info_label.setWordWrap(True)
+		#create buttons
+		self.checkout_button = QtGui.QPushButton('Checkout')
+		self.checkout_button.clicked.connect(self.checkout)
+		self.cancel_button = QtGui.QPushButton('Cancel')
+		self.cancel_button.clicked.connect(app.quit)    
+		
+		#create button layout
+		button_layout = QtGui.QHBoxLayout()
+		button_layout.setSpacing(2)
+		
+		button_layout.addWidget(self.checkout_button)
+		button_layout.addWidget(self.cancel_button)
+		
+		main_layout = QtGui.QVBoxLayout()
+		self.setLayout(main_layout)
+		main_layout.setSpacing(5)
+		main_layout.setMargin(6)
+		main_layout.addWidget(self.dept_tabs)
+		main_layout.addLayout(button_layout)
 
-        # Create action buttons
-        self.checkout_button = QtGui.QPushButton('Checkout')
-        self.cancel_button = QtGui.QPushButton('Cancel')
+		self.show()
 
-        # Create button layout
-        button_layout = QtGui.QHBoxLayout()
-        button_layout.setSpacing(2)
-
-        button_layout.addWidget(self.checkout_button)
-        button_layout.addWidget(self.cancel_button)
-
-        # Create main layout
-        main_layout = QtGui.QVBoxLayout()
-        self.setLayout(main_layout)
-        main_layout.setSpacing(5)
-        main_layout.setMargin(6)
-        #main_layout.addWidget(self.dept_tabs)
-        #main_layout.addWidget(self.asset_info_label)
-        main_layout.addLayout(button_layout)
-
-        # Change checkout context
-        #self.dept_tabs.currentChanged.connect(self.refresh)
-
-        # Connect the buttons
-        self.checkout_button.clicked.connect(self.checkout)
-        self.cancel_button.clicked.connect(self.cancel)
-       
-
-    def checkout(self):
-        """
-        Checks out the currently selected item
-        :return:
-        """
-        print('Checkout')
-
-    def cancel(self):
-        """
-        Cancels the checkout
-        :return:
-        """
-        self.close()
-
+	def checkout(self):
+		"""
+		Checks out the currently selected item
+		:return:
+		"""
+		print('Checkout')
 
 if __name__ == '__main__':
-    app = QtGui.QApplication(sys.argv)
-    window = CheckoutWindow()
-    window.show()
-    sys.exit(app.exec_())
+	app = QtGui.QApplication(sys.argv)
+	ex = checkoutWindow()
+	sys.exit(app.exec_())
